backend/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
Pillow>=10.4,<12.0
django-cloudinary-storage>=0.3,<1.0
cloudinary>=1.38,<2.0
